+++ src/Keys.php
@@ -1,14 +1,12 @@
 <?php
 
-namespace awssat\Visits;
+namespace Awssat\Visits;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
 
 class Keys
 {
-    public $prefix;
-    public $testing = '';
     public $modelName = false;
     public $id;
     public $visits;
@@ -16,16 +14,9 @@ class Keys
     public $instanceOfModel = false;
     public $tag;
 
-    /**
-     * Keys constructor.
-     * @param $subject
-     * @param $tag
-     */
     public function __construct($subject, $tag)
     {
         $this->modelName = $this->pluralModelName($subject);
-        $this->prefix = config('visits.redis_keys_prefix');
-        $this->testing = app()->environment('testing') ? 'testing:' : '';
         $this->primary = (new $subject)->getKeyName();
         $this->tag = $tag;
         $this->visits = $this->visits();
@@ -39,18 +30,14 @@ class Keys
 
     /**
      * Get cache key
-     *
-     * @return string
      */
     public function visits()
     {
-        return "{$this->prefix}:$this->testing" . $this->modelName . "_{$this->tag}";
+        return (app()->environment('testing') ? 'testing:' : '').$this->modelName."_{$this->tag}";
     }
 
     /**
-     * Get cache key
-     *
-     * @return string
+     * Get cache key for total values
      */
     public function visitsTotal()
     {
@@ -58,63 +45,47 @@ class Keys
     }
 
     /**
-     * @param $ip
-     * @return string
+     * ip key
      */
     public function ip($ip)
     {
-        return $this->visits . '_' .
-            snake_case("recorded_ips:" . ($this->instanceOfModel ? "{$this->id}:" : '') . $ip);
+        return $this->visits.'_'.Str::snake(
+            'recorded_ips:'.($this->instanceOfModel ? "{$this->id}:" : '') . $ip
+        );
     }
 
-
     /**
-     * @param $limit
-     * @param $isLow
-     * @return string
+     * list cache key
      */
     public function cache($limit = '*', $isLow = false)
     {
-        $key = $this->visits . "_lists";
+        $key = $this->visits.'_lists';
 
         if ($limit == '*') {
             return "{$key}:*";
         }
 
-        return "{$key}:" . ($isLow ? "low" : "top") . "{$limit}";
+        return "{$key}:".($isLow ? 'low' : 'top').$limit;
     }
 
     /**
-     * @param $period
-     * @return string
+     * period key
      */
     public function period($period)
     {
         return "{$this->visits}_{$period}";
     }
 
-    /**
-     * @param $relation
-     * @param $id
-     */
     public function append($relation, $id)
     {
         $this->visits .= "_{$relation}_{$id}";
     }
 
-    /**
-     * @param $subject
-     * @return string
-     */
     public function modelName($subject)
     {
         return strtolower(Str::singular(class_basename(get_class($subject))));
     }
 
-    /**
-     * @param $subject
-     * @return string
-     */
     public function pluralModelName($subject)
     {
         return strtolower(Str::plural(class_basename(is_string($subject) ? $subject : get_class($subject))));