+++ tests/Feature/EloquentPeriodsTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Awssat\Visits\Tests\Feature;
+
+
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+class EloquentPeriodsTest extends PeriodsTestCase
+{
+    use RefreshDatabase;
+
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        $this->app['config']['visits.engine'] = 'eloquent';
+        $this->connection = app(\Awssat\Visits\DataEngines\EloquentEngine::class)
+                            ->setPrefix($this->app['config']['visits.keys_prefix']);
+        include_once __DIR__.'/../../database/migrations/create_visits_table.php.stub';
+        (new \CreateVisitsTable())->up();
+    }
+}