+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All Notable changes to `laravel-visits` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 2.1.0
+
+- Rewrites huge part of the package to support multiple data engines.
+- Adds database's data engine support (Eloquent).
+
+
 ## 2.0.0
 
 - Global ignore feature (can be enabled from [config/visits.php](https://github.com/awssat/laravel-visits/blob/master/src/config/visits.php#L70))